Exercice_Django/requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14
django-filter>=24.2
django-cors-headers>=4.0